requirements-to-freeze.txt
requests
testutils
pyswagger
sentry_sdk
django
djangorestframework
markdown
django-filter
jsonpickle
drf_yasg
django-extensions
pylint
pylint-django
pytest-django
tqdm
flex
swagger_spec_validator
django-cors-headers
gunicorn
cathpy
biopython
celery
redis
psycopg2
django-redis
flower
python-decouple
